move attribute definitions to hir::attributes::*
